Removed cast warnings.
Former-commit-id: cba4eb43bfde51bf6a964e13fd795687a4684535 [formerly 8475d9d9eb29ddf69af805acb79ba924ac510bcc]
Former-commit-id: 1462c0c14c392df6648bf755f9b59cb83e1fc346 [formerly eecfe69184f527224f4b49b14d1bd5204f3f745c]
Former-commit-id: 2b735105fddffa78cb30a5e0b6e35cc700cf0b7e

+++ mrViewer/src/gui/mrvEDLGroup.cpp
@@ -504,7 +504,7 @@ int EDLGroup::handle( int event )
 	    }
 
             browser()->set_edl();
-            timeline()->value( pt );
+            timeline()->value( (double)pt );
             view()->seek( pt );
 	    delete _drag;
 	    _drag = NULL;
@@ -597,7 +597,7 @@ void EDLGroup::zoom( double z )
    tmin = tcur - int64_t( double(tlen) * pct );
 
 
-   if ( tmin < 0.0 ) tmin = 0.0;
+   if ( tmin < 0 ) tmin = 0;
 
    t->minimum( double(tmin) );
    t->maximum( double(tmax) );
@@ -686,14 +686,14 @@ void EDLGroup::merge( boost::int64_t frame )
     int64_t f = r->global_to_local( frame );
 
     int idx2 = idx;
-    if ( f < img->first_frame() + img->duration() / 2 )
+    if ( f < img->first_frame() + (int64_t) img->duration() / 2 )
     {
         idx -= 1;
     }
 
     idx2 = idx + 1;
 
-    if ( idx < 0 || idx2 >= r->images.size() ) return;
+    if ( idx < 0 || idx2 >= (int)r->images.size() ) return;
 
     mrv::media lm( r->images[idx] );
     if (!lm) return;
